+++ src/components/prompt-wizard/steps/index.ts
@@ -1,11 +1,9 @@
 export { TaskIntentStep } from "./TaskIntentStep";
 export { ContextStep } from "./ContextStep";
 export { ConstraintsStep } from "./ConstraintsStep";
-export { AudienceStep } from "./AudienceStep";
 export { OutputFormatStep } from "./OutputFormatStep";
 // Advanced steps (6-10)
 export { RoleStep } from "./RoleStep";
-export { ToneStep } from "./ToneStep";
 export { ReasoningStep } from "./ReasoningStep";
 export { SelfCheckStep } from "./SelfCheckStep";
 export { DisallowedStep } from "./DisallowedStep";